+++ Gemfile
@@ -7,6 +7,8 @@ ruby '3.3.0'
 # Bundle edge Rails instead: gem "rails", github: "rails/rails", branch: "main"
 gem 'rails', '~> 7.1.3'
 
+gem "cancancan"
+
 # Use postgresql as the database for Active Record
 gem 'pg', '~> 1.1'
