+++ app/models/role.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 class Role < ApplicationRecord
-  has_one :user
+  has_many :users
 end
