+++ app/serializers/movie_show_serializer.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
-
-class MovieShowSerializer < ActiveModel::Serializer
-  attributes :id, :language, :seat_count, :show_start_time, :show_end_time, :screen_no, :seat_type
+module Api
+  module V1
+    class MovieShowSerializer < ActiveModel::Serializer
+      attributes :id, :language, :seat_count, :show_start_time, :show_end_time, :screen_no, :seat_type
+    end
+  end
 end
