+++ app/serializers/user_serializer.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
 class UserSerializer < ActiveModel::Serializer
-  attributes :id, :name, :age, :phone_no
+  attributes :id, :name, :age, :phone_no, :role
+
+  def role
+    object.role.role_name if object.role
+  end
 end
