+++ config/routes.rb
@@ -6,15 +6,31 @@ Rails.application.routes.draw do
   # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
   # Can be used by load balancers and uptime monitors to verify that the app is live.
   get 'up' => 'rails/health#show', as: :rails_health_check
-  post '/login', to: 'users#login'
+  
+  
+  namespace :api do
+    namespace :v1 do
+      resources :users
+      resources :movie_shows
+      resources :movies
+      resources :theaters
+      resources :tickets
+      resources :movie_in_theaters
+      resources :roles
+      post '/login', to: 'users#login'
+      
+    
+    end
+  
+  end
 
   # Defines the root path route ("/")
   # root "posts#index"
-  resources :roles
-  resources :movies
-  resources :theaters
-  resources :movie_shows
-  resources :users
-  resources :movie_in_theaters
-  resources :tickets
+  
+  
+  
+  
+
+  
+  
 end
